PHP tests to see what is possible.
Almost got a working version of Sinatra requests... just need a
smigeon of routing to pull it all together. Albeit, it is a very
(very) simple implementation that lacks about 95% of the features
of Sinatra, but hey... it's a start in the right direction.

Just need to combine this with some code that I have at work, and
I think I should be good to go for a test-run of a simple web-api.