Minor Updates
I'm putting off a lot until I can get routing in tomorrow but I
decided to play with traits a bit so I added a little (tiny really)
JSON utility for returning JSON from inside the method handlers.

All it really does it serialize and set the appropriate header. But,
this should server as a (simple) guide for how to initially setup
template rendering (smarty and twig possibly??).