Combined sample tests to make first-working version
Yay! I have a sample get-request working. Tecnically, I could get
this thing setup in short-order for get and post requests but I
don't like the current route-matching. It's currently using standard
regular expressions and while that is nice... it's not very
developer-friendly (I have to remember that not everyone loves regular
expressions as much as I do).

So, got a basic working model done, now moving on to bigger and
better parsing/routing engine(s)!! :-]