Added support for HEAD, DELETE, and partial-PUT
The HEAD and DELETE are good to go, but I need to work on parsing the
body-data for PUT since PHP doesn't do that auto-magically for us.
Although I have read that there is a good function for doing just that.
Also updated the README.md with the latest updates in todo-completion